# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Veterinaria API",
        "description": "Documentación OpenAPI de la API de veterinaria.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor principal"
        }
    ],
    "paths": {
        "/api/pets": {
            "get": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Listar mascotas",
                "operationId": "fd9dbe447bbabd2069e0273aba5a573d",
                "responses": {
                    "200": {
                        "description": "Listado de mascotas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Pet"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Crear mascota",
                "operationId": "dd1b76666017d05ace0bcda388561641",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mascota creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pet"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pets/{pet}": {
            "get": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Ver mascota",
                "operationId": "9d5b229a4fde1ed73d595c1b1d3e6c42",
                "parameters": [
                    {
                        "name": "pet",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de mascota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mascota no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Actualizar mascota",
                "operationId": "920bc87be7d4cedd6ae23b869e954513",
                "parameters": [
                    {
                        "name": "pet",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mascota actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pet"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mascota no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Mascotas"
                ],
                "summary": "Eliminar mascota",
                "operationId": "b7b10b07ccfc1262bef72472a6def7d8",
                "parameters": [
                    {
                        "name": "pet",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Mascota eliminada"
                    },
                    "404": {
                        "description": "Mascota no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OwnerSummary": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "phone": {
                        "type": "string",
                        "example": "999888777",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Pet": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    },
                    "name": {
                        "type": "string",
                        "example": "Firulais"
                    },
                    "species": {
                        "type": "string",
                        "example": "Canino"
                    },
                    "race": {
                        "type": "string",
                        "example": "Mestizo"
                    },
                    "gender": {
                        "type": "string",
                        "example": "MACHO"
                    },
                    "temperament": {
                        "type": "string",
                        "example": "Tranquilo"
                    },
                    "reproductive_condition": {
                        "type": "string",
                        "example": "ENTERO"
                    },
                    "age": {
                        "type": "string",
                        "example": "3 años y 4 meses"
                    },
                    "status": {
                        "type": "string",
                        "example": "ADOPTADO"
                    },
                    "user": {
                        "$ref": "#/components/schemas/OwnerSummary"
                    }
                },
                "type": "object"
            },
...
            "StorePetRequest": {
                "required": [
                    "name",
                    "species",
                    "race",
                    "gender",
                    "temperament",
                    "reproductive_condition",
                    "color",
                    "years",
                    "months",
                    "status",
                    "user_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Firulais"
                    },
                    "species": {
                        "type": "string",
                        "example": "Canino"
                    },
                    "race": {
                        "type": "string",
                        "example": "Mestizo"
                    },
                    "gender": {
                        "type": "string",
                        "example": "MACHO"
                    },
                    "temperament": {
                        "type": "string",
                        "example": "Tranquilo"
                    },
                    "reproductive_condition": {
                        "type": "string",
                        "example": "ENTERO"
                    },
                    "color": {
                        "type": "string",
                        "example": "Marrón"
                    },
                    "years": {
                        "type": "integer",
                        "example": 3
                    },
                    "months": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "example": "ADOPTADO"
                    },
                    "user_id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    }
                },
                "type": "object"
            },
            "UpdatePetRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Firulais"
                    },
                    "temperament": {
                        "type": "string",
                        "example": "Juguetón"
                    },
                    "reproductive_condition": {
                        "type": "string",
                        "example": "CASTRADO"
                    },
                    "color": {
                        "type": "string",
                        "example": "Negro"
                    },
                    "years": {
                        "type": "integer",
                        "example": 4
                    },
                    "months": {
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "type": "string",
                        "example": "ADOPTADO"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "No autenticado."
                    },
                    "status": {
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Los datos enviados no son válidos."
                    },
                    "status": {
                        "type": "integer",
                        "example": 422
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "El campo email es obligatorio."
                                ]
                            },
                            "password": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "El campo password es obligatorio."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación JWT usando Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}